requirements.txt
asgiref==3.8.1
Django==5.0.4
djangorestframework==3.15.1
sqlparse==0.5.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
pillow==10.3.0
psycopg2-binary==2.9.9